requirements.txt
docker==3.7.2
PyYAML==4.2b1
